requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
line-bot-sdk>=3.12.0
openai>=1.65.0
# PostgreSQL 驅動（當設定 DATABASE_URL 時，app/db.py 以 psycopg 直連）
psycopg[binary]>=3.2.0
python-dotenv>=1.0.1
httpx>=0.27.0
google-auth>=2.35.0
google-cloud-storage>=2.18.0
redis>=5.0.0
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48b0
